clk: MediaTek: add clock driver for MT7629 SoC.
This patch adds clock driver for MediaTek SoC and divides it into two parts:

- Shared functions: a common driver which contains the general operations for
plls, muxes, dividers and clock gates so that we can reuse it for other SoCs.

- Specific SoC drivers: the group of structures which describe the hardware
configurations.

And we take MT7629 as an example to demonstrate how to implement drivers if
any other MediaTek chips would like to use it.

Signed-off-by: Ryder Lee <ryder.lee@mediatek.com>