pinctrl: MediaTek: add pinctrl driver for MT7629 SoC
Add support for pinctrl on MT7629 SoC. The IO core found on the SoC
has the registers for pinctrl, pinconf and gpio mixed up in the same
register range. Hence the driver also implements the gpio functionality
through UCLASS_GPIO.

We create pinctrl-mtk-common.c for common part as there might be other
chips that use the same binding and driver, then being a little more
abstract could help in the long run.

Signed-off-by: Ryder Lee <ryder.lee@mediatek.com>